+++ memcheck/tests.R
@@ -1,4 +1,3 @@
 # Code to be run with
 #   R -d "valgrind --tool=memcheck --leak-check=full --error-exitcode=1" --vanilla < memcheck/thisfile.R
-library("TreeTools")
-devtools::test()
+testthat::test_local()